+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "editor.formatOnSave": true,
+    "editor.wordWrap": "on",
+    "editor.wordWrapColumn": 120,
+    "prettier.singleQuote": true
+}
