+++ App.js
@@ -1,21 +1,11 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
-import Network from './screen/network';
+import React from "react";
+
+import HomeScreen from "./screen/HomeScreen";
 
 export default function App() {
   return (
-    <View style={styles.container}>
-      <Text>Hello this is testing</Text>
-      <StatusBar style="auto" />
-    </View>
+    <HomeScreen />
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
+
