+++ comps/Header.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { View, StyleSheet, Text } from 'react-native'
+
+const Header = () => {
+    return (
+        <View>
+            <Text>Label</Text>
+        </View>
+    )
+}
+
+export default Header
